+++ fractal_server/app/schemas/v2/job.py
@@ -13,8 +13,9 @@ from pydantic.types import AwareDatetime
 from pydantic.types import StrictStr
 
 from .._filter_validators import validate_attribute_filters
+from .._validators import cant_set_none
+from .._validators import NonEmptyString
 from .._validators import root_validate_dict_keys
-from .._validators import valstr
 from .dumps import DatasetDumpV2
 from .dumps import ProjectDumpV2
 from .dumps import WorkflowDumpV2
@@ -49,14 +50,17 @@ class JobCreateV2(BaseModel):
     first_task_index: Optional[int] = None
     last_task_index: Optional[int] = None
     slurm_account: Optional[StrictStr] = None
-    worker_init: Optional[str] = None
+    worker_init: Optional[NonEmptyString] = None
 
     attribute_filters: AttributeFiltersType = Field(default_factory=dict)
 
     # Validators
-    _worker_init = field_validator("worker_init")(
-        classmethod(valstr("worker_init"))
-    )
+
+    @field_validator("worker_init")
+    @classmethod
+    def _cant_set_none(cls, v):
+        return cant_set_none(v)
+
     _dict_keys = model_validator(mode="before")(
         classmethod(root_validate_dict_keys)
     )