+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * App:
     * Obfuscate sensitive information from settings using `SecretStr` (\#2333).
+* API:
+    * Remove `valstr` validator and introduce `NonEmptyString` in schemas (\#2352).
 
 
 # 2.14.0a0, 2.14.0a1, 2.14.0a2